# 全域md文档/02-开发与设计文档/openapi.yaml
openapi: 3.0.3
...
info:
  title: 课堂管理系统 - 教师端 API
  version: 1.0.0
  description: |
    教师端 Node 进程（Express 4.22.2）对外暴露的接口，服务对象是同机运行的大屏看板浏览器。

    ## 架构位置

    ```
    学生机 Electron ──MQTT WS 1888/ws──► SIoT2 ──MQTT TCP 1883──► 教师端 Node
                                                                      │
                                                          唯一聚合点 + SQLite 权威源
                                                                      │
                                                    ┌─────────────────┴──────────────┐
                                                    │  本文件定义的 HTTP + SSE 契约     │
                                                    └─────────────────┬──────────────┘
                                                                      ▼
                                                              大屏看板浏览器
    ```

    ## 统一响应包装

    除文件下载（`/exports/{exportId}/files/{fileName}`）与 SSE 流（`/dashboard/stream`）外，
    **所有**端点返回统一信封：

    ```json
    { "code": 0, "data": { }, "message": "" }
    ```

    - `code` = 0 表示成功；非 0 表示业务错误，此时 `data` 为 null。
    - `errorCode` 仅在失败时出现，取值对齐 04 文档错误码表（如 `E-VAL-01`）。
    - HTTP 状态码同时正确设置（不做"200 包一切"），便于前端用拦截器统一处理。

    ## 时间与标识约定

    - 所有时间戳为 **Unix 毫秒整数**（与 MQTT 信封 `ts` 一致），前端负责本地化显示。
    - `seat` 为两位字符串（`"01"`–`"50"`），**不是数字**——`"07"` 与 `7` 不等价，禁止隐式转换。
    - `msgId` / `sessionId` / `taskId` 等标识为字符串，不假设为 UUID 格式。

    ## 认证（部署形态决定）

    默认部署下 Node 仅监听 `127.0.0.1`，看板与服务同机，无需认证。
...

  contact:
...
servers:
...
security:
  - TeacherToken: []
paths:
...
  /commands/shutdown:
    post:
      tags: [Command]
      summary: 下发关机指令
      description: |
        **本端点是全系统唯一的破坏性操作，服务端必须做三重校验后才广播**（SPEC §10、§11）：

        1. **阶段闸门**：`phase` 必须为 `return` 或 `closed`。否则 `409` + `E-OFF-01`。
           防止课中误触让全班关机。
        2. **归还闸门**：`force=false` 时要求全部小组 `allReturned=true`，
           否则 `409` + `E-RETURN-01` 并在 `data.pendingSeats` 返回未归还座位，
           看板据此提���教师。`force=true` 由教师显式确认后跳过此闸门（AC-4）。
        3. **HMAC 签名**：服务端用 `sharedSecret` 计算
           `sig = HMAC_SHA256(secret, sessionId + "|shutdown|" + ts)` 写入 MQTT 载荷。
           学生端校验 `sig` 有效、`phase ∈ {return, closed}` 且 `|now - ts| < 60s` 才执行关机。
           这三条共同防止学生用 MQTTX 伪造或重放关机报文。

        **离线机兜底**（PRD R-7）：`data.offlineSeats` 返回广播时已离线的座位。
        这些机器收不到指令，看板必须标注"尚未关机"清单交教师现场处理；
        若它们随后重连，会经 hello→sync 得知已进入关机阶段并由教师端补发。
        MVP 不承诺 100% 覆盖离线机。
      operationId: sendShutdown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
                  description: 跳过「全部归还」闸门（教师强制关机，需 UI 二次确认）
                reason:
                  type: string
                  maxLength: 200
                  description: 强制关机原因，写入事件日志备查
            example:
              force: false
      responses:
        '200':
          description: 指令已广播
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShutdownResult'
              example:
                code: 0
                message: ''
                data:
                  broadcast:
                    delivered: true
                    msgId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                    topic: siot/ict_cmd
                  shutdownDelaySec: 10
                  targetOnline: 48
                  offlineSeats: ['13', '29']
                  pendingSeats: []
        '409':
          description: 闸门未通过（阶段不符 或 尚有未归还）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                phaseGate:
                  summary: 阶段闸门拦截
                  value:
                    code: 40903
                    message: 仅在课后归还阶段可下发关机指令，当前阶段 task
                    errorCode: E-OFF-01
                    data: null
                returnGate:
                  summary: 归还闸门拦截（可用 force 跳过）
                  value:
                    code: 40904
                    message: 尚有 3 个座位未确认归还，如需强制关机请设 force=true
                    errorCode: E-RETURN-01
                    data:
                      pendingSeats: ['13', '22', '29']
        '503':
          $ref: '#/components/responses/MqttUnavailable'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commands/reset:
    post:
      tags: [Command]
      summary: 复位指定座位的登记
      description: |
        教师代操作，用于学生填错信息重填。清除该座位 `t_student` 登记态与本课器材领取记录，
        并向学生机广播 `{action:"reset", seat}`。学生端仅当 `payload.seat` 等于本机座位时响应。
      operationId: sendReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seat]
              properties:
                seat:
                  $ref: '#/components/schemas/Seat'
            example:
              seat: '07'
      responses:
        '200':
          description: 已复位并广播
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          seat:
                            $ref: '#/components/schemas/Seat'
                          broadcast:
                            $ref: '#/components/schemas/BroadcastResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/MqttUnavailable'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================================================
  # Dashboard
  # ==========================================================================
...
components:
  securitySchemes:
    TeacherToken:
      type: apiKey
      in: header
      name: X-Teacher-Token
      description: |
        **默认部署（监听 127.0.0.1）下不启用**——看板与服务同机，无网络暴露面。

        **绑定到局域网地址时必须启用**：否则任何学生只要知道教师机 IP 与端口，
        就能直接 `POST /api/v1/commands/shutdown` 让全班关机——这比伪造 MQTT 报文更容易，
        因为 HTTP 端点没有 HMAC 保护。令牌由教师端启动时生成并打印在控制台，
        看板首次访问时填入并存 `localStorage`。
...
  responses:
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: 40001
            message: seat 必须为两位数字字符串，如 "07"
            errorCode: E-VAL-01
            data: null
...
    Unauthorized:
      description: 缺少或错误的 `X-Teacher-Token`（仅局域网部署形态会出现）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: 40100
            message: 无效的教师端令牌
            errorCode: E-AUTH-01
            data: null
    MqttUnavailable:
      description: |
        MQTT 未连接，指令无法送达。**本地状态变更仍已持久化**，
        看板应提示"指令未送达，请检查 SIoT2 连接后重试"。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: 50301
            message: SIoT2 连接不可用，指令未送达
            errorCode: E-CONN-01
            data: null
  schemas:
...
    ApiEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0 = 成功；非 0 = 业务错误码
          example: 0
        message:
          type: string
          description: 成功时为空串；失败时为可直接展示给教师的中文描述
          example: ''
        data:
          nullable: true
          description: 业务数据；失败时为 null
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: |
            业务错误码。前两位对应 HTTP 语义，后三位为细分：
            `40001` 参数校验 / `40100` 认证 / `40400` 不存在 /
            `40901` 阶段非法 / `40902` 会话冲突 / `40903` 关机阶段闸门 /
            `40904` 归还闸门 / `50000` 服务内部 / `50301` MQTT 不可用
          example: 40001
        message:
          type: string
          example: seat 必须为两位数字字符串
        errorCode:
          type: string
          description: 对齐 04 文档错误码表，便于跨端一致处理
          enum:
            - E-CONN-01
            - E-AUTH-01
            - E-DUP-01
            - E-VAL-01
            - E-OFF-01
            - E-SESSION-01
            - E-SESSION-02
            - E-PHASE-01
            - E-RETURN-01
            - E-NOTFOUND
            - E-INTERNAL
          example: E-VAL-01
        data:
          nullable: true
          description: 通常为 null；闸门类错误会带辅助信息（如 `pendingSeats`）
...
    Seat:
      type: string
      pattern: '^[0-9]{2}$'
      description: 座位号，两位数字字符串（`"01"`–`"50"`）。**不是整数**，禁止隐式转换。
      example: '07'
...
    BroadcastResult:
      type: object
      description: MQTT 广播结果。`delivered=false` 表示本地已处理但指令未送达。
      required: [delivered]
      properties:
        delivered:
          type: boolean
          example: true
        msgId:
          type: string
          nullable: true
          description: 与 MQTT 信封 `msgId` 一致，便于在 SIoT Web 页对账
          example: 550e8400-e29b-41d4-a716-446655440000
        topic:
          type: string
          nullable: true
          example: siot/ict_cmd
        error:
          type: string
          nullable: true
          description: delivered=false 时的原因
          example: null
...
    ShutdownResult:
      type: object
      properties:
        broadcast:
          $ref: '#/components/schemas/BroadcastResult'
        shutdownDelaySec:
          type: integer
          example: 10
        targetOnline:
          type: integer
          description: 广播时在线的学生机数
          example: 48
        offlineSeats:
          type: array
          description: 广播时已离线、收不到指令的座位。看板必须标注"尚未关机"（PRD R-7）。
          items:
            $ref: '#/components/schemas/Seat'
        pendingSeats:
          type: array
          description: force=true 时被跳过的未归还座位，留档备查
          items:
            $ref: '#/components/schemas/Seat'

    # ---------- Dashboard ----------
...